moving wndcharm dependencies in their own module and changin all iteritems references to items to make python 2 and 3 compliant
